// apps/web-shared/src/model/trains/TrainType.ts
enum TrainType {
  Bike = 'bike',
  Boat = 'boat',
  Bus = 'bus',
  Car = 'car',
  CarWithTrailer = 'car-with-trailer',
  Motorcycle = 'motorcycle',
  Plane = 'plane',
  TrainDiesel = 'train-diesel',
  TrainElectric = 'train-electric',
  TrainSteam = 'train-steam',
  TrainMetro = 'train-metro',
  Tram = 'tram',
  Truck = 'truck',
  TruckWithTrailer = 'truck-with-trailer',
}

export default TrainType;
export { TrainType };